MATLAB Folder and Video to Images script
So I wrote this script that will convert a video to its pictures by
frame. All the pre-work that needs to be done is that you need to have
a VidtPic directory in your MATLAB folder. The video you want to
convert to images also needs to be in that directory. The script will
then make a subdirectory for the pictures and it will move the video
from VidtPic to the new directory so that all the contents are
consolidated.